% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_query.R
\name{apply_on_each_term}
\alias{apply_on_each_term}
\title{Apply a Function to Each Term of the Query}
\usage{
apply_on_each_term(query, fun)
}
\arguments{
\item{query}{A search query as a string.}

\item{fun}{A function to apply to each term.}
}
\value{
A new query as a string with each term transformed by the provided function.
}
\description{
This function applies a given function to each term in a query.
}
\examples{
apply_on_each_term("[term1] AND [term2]", toupper)
}
