% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{check_write_access}
\alias{check_write_access}
\title{Check Write Access}
\usage{
check_write_access(path)
}
\arguments{
\item{path}{A file path.}
}
\description{
Check if you can write on the provided path or raise an error otherwise.
}
\examples{
check_write_access(tempfile())
}
