% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(url)
}
\arguments{
\item{url}{A search URL.}
}
\value{
A list of data from medRxiv/bioRxiv.
}
\description{
This function retrieves a list of data from medRxiv/bioRxiv given a search URL.
}
\examples{
url <- "https://www.medrxiv.org/search/abstract_title\%3ACOVID-19\%20abstract_title_flags\%3Amatch-all\%20limit_from\%3A2020-01-01\%20limit_to\%3A2021-01-01\%20numresults\%3A75\%20sort\%3Apublication-date\%20direction\%3Adescending\%20format_result\%3Acondensed"
get_data(url)
}
