% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_query.R
\name{get_max_group_level}
\alias{get_max_group_level}
\title{Get the Maximum Group Level of a Query}
\usage{
get_max_group_level(query)
}
\arguments{
\item{query}{A search query as a string.}
}
\value{
The maximum group level as an integer.
}
\description{
This function returns the maximum group level of a query.
}
\examples{
get_max_group_level("term1 AND (term2 OR (term3 AND term4))")
}
