% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{get_numeric_month_by_string}
\alias{get_numeric_month_by_string}
\title{Get Numeric Month by String}
\usage{
get_numeric_month_by_string(string, fallback_month = "01")
}
\arguments{
\item{string}{A month string representation (e.g., "jan", "January", "Jan", "Feb", "December").}

\item{fallback_month}{A fallback month string to be returned on any error, default is "01".}
}
\value{
A month numeric representation as a string (e.g., "jan" -> "01").
}
\description{
Get a numeric month representation given a month string representation.
}
\examples{
get_numeric_month_by_string("Jan")
get_numeric_month_by_string("December", "12")
}
