% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_page_count.R
\name{get_page_count}
\alias{get_page_count}
\title{Get Page Count}
\usage{
get_page_count(response)
}
\arguments{
\item{response}{A httr2 response object.}
}
\value{
The page count.
}
\description{
Retrieve the page count from the search results.
}
\examples{
page_count <- get_page_count(response)
}
