% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paper.R
\name{get_paper}
\alias{get_paper}
\title{Get Paper}
\usage{
get_paper(paper_metadata)
}
\arguments{
\item{paper_metadata}{A list containing paper metadata.}
}
\value{
A Paper object.
}
\description{
This function creates a Paper object from metadata.
}
\examples{
paper_metadata <- list(
  title = "COVID-19 and mental health",
  abstract = "Abstract content",
  authors = "Author1; Author2",
  date = "2020-01-01",
  doi = "10.1101/2020.01.01.20016730"
)
get_paper(paper_metadata)
}
