% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paper_metadata.R
\name{get_paper_metadata}
\alias{get_paper_metadata}
\title{Get Paper Metadata}
\usage{
get_paper_metadata(doi, database)
}
\arguments{
\item{doi}{The paper DOI.}

\item{database}{The database name (either "medRxiv" or "bioRxiv").}
}
\value{
A list containing the medRxiv/bioRxiv paper metadata, or NULL if no metadata is available.
}
\description{
This function retrieves a paper's metadata given its DOI.
}
\examples{
doi <- "10.1101/2020.01.01.20016730"
get_paper_metadata(doi, "medRxiv")
}
