% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_query.R
\name{get_query_tree}
\alias{get_query_tree}
\title{Get the Tree of a Query}
\usage{
get_query_tree(query, parent = NULL)
}
\arguments{
\item{query}{A search query as a string.}

\item{parent}{A parent node as a list (default is NULL).}
}
\value{
The query tree as a list.
}
\description{
This function returns a tree representation of a query.
}
\examples{
get_query_tree("[term A] OR [term B] AND ([term C] OR [term D])")
}
