% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result.R
\name{get_result}
\alias{get_result}
\title{Get Result}
\usage{
get_result(url)
}
\arguments{
\item{url}{A URL to search for results.}
}
\value{
A page from medRxiv/bioRxiv databases.
}
\description{
This function returns the results from medRxiv/bioRxiv databases using the provided search parameters.
}
\examples{
url <- "https://www.medrxiv.org/search/abstract_title\%3ACOVID-19\%20abstract_title_flags\%3Amatch-all\%20limit_from\%3A2020-01-01\%20limit_to\%3A2021-01-01\%20numresults\%3A75\%20sort\%3Apublication-date\%20direction\%3Adescending\%20format_result\%3Acondensed"
get_result(url)
}
