% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result_page_data.R
\name{get_result_page_data}
\alias{get_result_page_data}
\title{Get Result Page Data}
\usage{
get_result_page_data(result_page)
}
\arguments{
\item{result_page}{A result page extracted from medRxiv/bioRxiv databases.}
}
\value{
A list containing paper DOIs, total papers, and next page URL information.
}
\description{
This function extracts result data from the given result page.
}
\examples{
result_page <- get_result("https://www.medrxiv.org/search/abstract_title\%3ACOVID-19\%20abstract_title_flags\%3Amatch-all\%20limit_from\%3A2020-01-01\%20limit_to\%3A2021-01-01\%20numresults\%3A75\%20sort\%3Apublication-date\%20direction\%3Adescending\%20format_result\%3Acondensed")
get_result_page_data(result_page)
}
