% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_httr2.R
\name{get_results}
\alias{get_results}
\title{Get Results}
\usage{
get_results(response)
}
\arguments{
\item{response}{A httr2 response object.}
}
\value{
A tibble of search results.
}
\description{
Extract the results from the search response.
}
\examples{
results <- get_results(response)
}
