% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_search_urls.R
\name{get_search_urls}
\alias{get_search_urls}
\title{Get Search URLs}
\usage{
get_search_urls(search, database)
}
\arguments{
\item{search}{A search instance containing query and date range.}

\item{database}{The database name (either "medRxiv" or "bioRxiv").}
}
\value{
A list of URLs to be used to retrieve data from medRxiv/bioRxiv databases.
}
\description{
This function returns the URLs to be used to retrieve data from the medRxiv/bioRxiv databases.
}
\examples{
search <- list(query = "COVID-19", since = as.Date("2020-01-01"), until = as.Date("2021-01-01"))
if(FALSE){
get_search_urls(search, "medRxiv")
}
}
