% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{logging_initialize}
\alias{logging_initialize}
\title{Logging Initialize}
\usage{
logging_initialize(verbose = FALSE)
}
\arguments{
\item{verbose}{A logical value indicating if the logging needs to be verbose (default is FALSE).}
}
\description{
Initialize logging. If verbose mode is TRUE, the logging will be initialized on DEBUG mode.
Otherwise, INFO mode will be used.
}
\examples{
logging_initialize(verbose = TRUE)
}
