% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_query.R
\name{replace_search_term_enclosures}
\alias{replace_search_term_enclosures}
\title{Replace Search Term Enclosures}
\usage{
replace_search_term_enclosures(
  query,
  open_replacement,
  close_replacement,
  only_on_wildcards = FALSE
)
}
\arguments{
\item{query}{A search query as a string.}

\item{open_replacement}{A string to replace the opening enclosure.}

\item{close_replacement}{A string to replace the closing enclosure.}

\item{only_on_wildcards}{A logical value indicating whether to replace enclosures only on wildcard terms (default is FALSE).}
}
\value{
A transformed query as a string.
}
\description{
This function replaces the enclosures of search terms in a query.
}
\examples{
replace_search_term_enclosures("[term1] AND [term2]", "(", ")", TRUE)
}
