% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run Search}
\usage{
run(search, database)
}
\arguments{
\item{search}{A search instance.}

\item{database}{The database name (medRxiv or bioRxiv).}
}
\value{
A tibble of search results.
}
\description{
This function fetches papers from medRxiv/bioRxiv databases using the provided search parameters.
The collected papers are added to the provided search instance.
}
\examples{
search <- list(query = "COVID-19", since = as.Date("2020-01-01"), until = as.Date("2021-01-01"))
run(search, "medRxiv")


Run the Search

Run the search on medRxiv or bioRxiv database.

search <- list(query = "COVID-19", since = as.Date("2020-01-01"), until = as.Date("2021-01-01"))
results <- run(search, "medRxiv")
}
