% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_search.R
\name{Search}
\alias{Search}
\title{A class to represent a Search}
\description{
A class to represent a Search

A class to represent a Search
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{query}}{The query used to fetch the papers}

\item{\code{since}}{The lower bound (inclusive) date of search}

\item{\code{until}}{The upper bound (inclusive) date of search}

\item{\code{limit}}{The max number of papers that can be returned in the search}

\item{\code{limit_per_database}}{The max number of papers that can be returned in the search for each database}

\item{\code{processed_at}}{The datetime when the search was performed}

\item{\code{databases}}{List of databases where the search should be performed}

\item{\code{publication_types}}{List of publication types to filter when searching}

\item{\code{papers}}{A list of papers already collected}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Search-new}{\code{Search$new()}}
\item \href{#method-Search-get_paper_key}{\code{Search$get_paper_key()}}
\item \href{#method-Search-get_publication_key}{\code{Search$get_publication_key()}}
\item \href{#method-Search-add_paper}{\code{Search$add_paper()}}
\item \href{#method-Search-get_paper}{\code{Search$get_paper()}}
\item \href{#method-Search-get_publication}{\code{Search$get_publication()}}
\item \href{#method-Search-remove_paper}{\code{Search$remove_paper()}}
\item \href{#method-Search-merge_duplications}{\code{Search$merge_duplications()}}
\item \href{#method-Search-reached_its_limit}{\code{Search$reached_its_limit()}}
\item \href{#method-Search-clone}{\code{Search$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Search-new"></a>}}
\if{latex}{\out{\hypertarget{method-Search-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Search$new(
  query,
  since = NULL,
  until = NULL,
  limit = NULL,
  limit_per_database = NULL,
  processed_at = NULL,
  databases = NULL,
  publication_types = NULL,
  papers = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{The query used to fetch the papers}

\item{\code{since}}{The lower bound (inclusive) date of search}

\item{\code{until}}{The upper bound (inclusive) date of search}

\item{\code{limit}}{The max number of papers that can be returned in the search}

\item{\code{limit_per_database}}{The max number of papers that can be returned in the search for each database}

\item{\code{processed_at}}{The datetime when the search was performed}

\item{\code{databases}}{List of databases where the search should be performed}

\item{\code{publication_types}}{List of publication types to filter when searching}

\item{\code{papers}}{A list of papers already collected}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Search} object.
Get a paper key
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Search-get_paper_key"></a>}}
\if{latex}{\out{\hypertarget{method-Search-get_paper_key}{}}}
\subsection{Method \code{get_paper_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Search$get_paper_key(paper_title, publication_date, paper_doi = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paper_title}}{Title of the paper}

\item{\code{publication_date}}{Date of publication}

\item{\code{paper_doi}}{DOI of the paper}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The paper key
Get a publication key
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Search-get_publication_key"></a>}}
\if{latex}{\out{\hypertarget{method-Search-get_publication_key}{}}}
\subsection{Method \code{get_publication_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Search$get_publication_key(
  publication_title,
  publication_issn = NULL,
  publication_isbn = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{publication_title}}{Title of the publication}

\item{\code{publication_issn}}{ISSN of the publication}

\item{\code{publication_isbn}}{ISBN of the publication}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The publication key
Add a paper to the search
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Search-add_paper"></a>}}
\if{latex}{\out{\hypertarget{method-Search-add_paper}{}}}
\subsection{Method \code{add_paper()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Search$add_paper(paper)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paper}}{A paper object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
Get a paper from the search
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Search-get_paper"></a>}}
\if{latex}{\out{\hypertarget{method-Search-get_paper}{}}}
\subsection{Method \code{get_paper()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Search$get_paper(paper_title = NULL, publication_date = NULL, paper_doi = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paper_title}}{Title of the paper}

\item{\code{publication_date}}{Date of publication}

\item{\code{paper_doi}}{DOI of the paper}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The paper object
Get a publication from the search
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Search-get_publication"></a>}}
\if{latex}{\out{\hypertarget{method-Search-get_publication}{}}}
\subsection{Method \code{get_publication()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Search$get_publication(title = NULL, issn = NULL, isbn = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{Title of the publication}

\item{\code{issn}}{ISSN of the publication}

\item{\code{isbn}}{ISBN of the publication}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The publication object
Remove a paper from the search
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Search-remove_paper"></a>}}
\if{latex}{\out{\hypertarget{method-Search-remove_paper}{}}}
\subsection{Method \code{remove_paper()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Search$remove_paper(paper)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paper}}{A paper object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
Merge duplicate papers based on similarity threshold
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Search-merge_duplications"></a>}}
\if{latex}{\out{\hypertarget{method-Search-merge_duplications}{}}}
\subsection{Method \code{merge_duplications()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Search$merge_duplications(similarity_threshold = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{similarity_threshold}}{The similarity threshold to consider two papers as duplicates}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
Check if the limit of papers is reached for a given database
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Search-reached_its_limit"></a>}}
\if{latex}{\out{\hypertarget{method-Search-reached_its_limit}{}}}
\subsection{Method \code{reached_its_limit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Search$reached_its_limit(database)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database}}{The database to check}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if the limit is reached, FALSE otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Search-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Search-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Search$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
