% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_search.R
\name{Search_fromDict}
\alias{Search_fromDict}
\title{A method that returns a Search instance based on the provided dict object}
\arguments{
\item{search_dict}{A dict that represents a Search instance}
}
\value{
A Search instance based on the provided dict object
}
\description{
A method that returns a Search instance based on the provided dict object
}
