% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_search.R
\name{Search_toDict}
\alias{Search_toDict}
\title{A method that returns a dict object based on the provided Search instance}
\arguments{
\item{search}{A Search instance}
}
\value{
A dict that represents a Search instance
}
\description{
A method that returns a dict object based on the provided Search instance
}
