% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{ThreadSafeSingletonMetaclass}
\alias{ThreadSafeSingletonMetaclass}
\title{Thread Safe Singleton Metaclass}
\value{
An instance of the class.
}
\description{
A thread-safe singleton metaclass.
}
\examples{
singleton_class <- ThreadSafeSingletonMetaclass$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ThreadSafeSingletonMetaclass-new}{\code{ThreadSafeSingletonMetaclass$new()}}
\item \href{#method-ThreadSafeSingletonMetaclass-call}{\code{ThreadSafeSingletonMetaclass$call()}}
\item \href{#method-ThreadSafeSingletonMetaclass-clone}{\code{ThreadSafeSingletonMetaclass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThreadSafeSingletonMetaclass-new"></a>}}
\if{latex}{\out{\hypertarget{method-ThreadSafeSingletonMetaclass-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThreadSafeSingletonMetaclass$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThreadSafeSingletonMetaclass-call"></a>}}
\if{latex}{\out{\hypertarget{method-ThreadSafeSingletonMetaclass-call}{}}}
\subsection{Method \code{call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThreadSafeSingletonMetaclass$call(cls, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cls}}{The class to be instantiated as a singleton.}

\item{\code{...}}{Additional arguments passed to the class constructor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThreadSafeSingletonMetaclass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ThreadSafeSingletonMetaclass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThreadSafeSingletonMetaclass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
