% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_common.R
\name{try_success}
\alias{try_success}
\title{Try Success}
\usage{
try_success(fun, attempts = 1, pre_delay = 0, next_try_delay = 3)
}
\arguments{
\item{fun}{A function to be tried.}

\item{attempts}{The number of attempts (default is 1).}

\item{pre_delay}{The delay before each function attempt in seconds (default is 0).}

\item{next_try_delay}{The delay between function attempts in seconds (default is 3).}
}
\value{
The returned value of the function or NULL if the function raises an exception in all attempts.
}
\description{
Try to execute a function and repeat this execution if it raises any exception.
This function will try N times to succeed, by provided number of attempts.
}
\examples{
try_success(function() { stop("error") }, attempts = 3)
}
